+++ kifi-backend/modules/shoebox/app/com/keepit/common/store/S3ImageStore.scala
@@ -224,23 +224,11 @@ class S3ImageStoreImpl @Inject() (
     }
   }
 
-  private def forceRGB(image: BufferedImage): BufferedImage = {
-    // This forces an image to use RGB and to use white as the transparency color if the source image supports it
-    // However, this can still fail on different color modes, especially from images explicitly saved as CMYK from
-    // Adobe software. The true solution is to use a full featured image processor, like imagemagick.
-    val imageRGB = new BufferedImage(image.getWidth, image.getHeight, BufferedImage.TYPE_INT_RGB)
-    val g = imageRGB.createGraphics()
-    g.setColor(Color.WHITE)
-    g.fillRect(0,0,image.getWidth,image.getHeight)
-    g.drawRenderedImage(image, null)
-    g.dispose()
-    imageRGB
-  }
   def readImage(file: File): BufferedImage = {
-    forceRGB(ImageIO.read(file))
+    ImageUtils.forceRGB(ImageIO.read(file))
   }
   def readImage(is: InputStream): BufferedImage = {
-    forceRGB(ImageIO.read(is))
+    ImageUtils.forceRGB(ImageIO.read(is))
   }
 
   def uploadTemporaryPicture(file: File): Try[(String, String)] = Try {
